# backend/src/tasks/dto/update-task.dto.ts
import { IsEnum, IsOptional, IsString } from 'class-validator';
import { TaskStatus } from '@prisma/client'; // Import TaskStatus enum

export class UpdateTaskDto {
  @IsOptional()
  @IsString()
  title?: string;

  @IsOptional()
  @IsString()
  description?: string;

  @IsOptional()
  @IsEnum(TaskStatus) // Validate that the status is one of the enum values
  status?: TaskStatus;
}