# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Task Management API
  description: API documentation for Task Management Service
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server

paths:
  /tasks:
    get:
      summary: Retrieve all tasks
      tags:
        - Tasks
      responses:
        200:
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

    post:
      summary: Create a new task
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskDto'
      responses:
        201:
          description: The created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

  /tasks/{id}:
    get:
      summary: Retrieve a task by ID
      tags:
        - Tasks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the task to retrieve
      responses:
        200:
          description: The task with the specified ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        404:
          description: Task not found

    patch:
      summary: Update a task by ID
      tags:
        - Tasks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the task to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskDto'
      responses:
        200:
          description: The updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        404:
          description: Task not found

    delete:
      summary: Delete a task by ID
      tags:
        - Tasks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the task to delete
      responses:
        200:
          description: The deleted task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        404:
          description: Task not found

components:
  schemas:
    CreateTaskDto:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
          description: The title of the task
          example: 'Complete documentation'
        description:
          type: string
          description: The description of the task
          example: 'Finish the API documentation by end of the day'

    Task:
      type: object
      properties:
        id:
          type: integer
          description: The ID of the task
          example: 1
        title:
          type: string
          description: The title of the task
          example: 'Complete documentation'
        description:
          type: string
          description: The description of the task
          example: 'Finish the API documentation by end of the day'
        status:
          type: string
          description: The status of the task
          example: 'pending'
        createdAt:
          type: string
          format: date-time
          description: The creation date of the task
        updatedAt:
          type: string
          format: date-time
          description: The last update date of the task

    UpdateTaskDto:
      type: object
      properties:
        title:
          type: string
          description: The title of the task
          example: 'Updated title'
        description:
          type: string
          description: The description of the task
          example: 'Updated description of the task'
        status:
          type: string
          enum: [pending, completed, cancelled]
          description: The status of the task
          example: 'completed'
      required:
        - status